error mimetype files

+++ assets/js/importdxf.js
@@ -4,8 +4,8 @@ document.getElementById("form-dxf").addEventListener("submit", (e) => {
         e.preventDefault();
         showSpinner();
         uploadDXF();
-    }
-})  
+    };
+})  ;
 
 // ----------- Show file name + size -----------
 const file = document.getElementById('file');
@@ -15,62 +15,69 @@ file.addEventListener('change', (e) => {
     const fileSize = (size / 1000).toFixed(2);
     const fileNameAndSize = `${fileName} - ${fileSize} KB`;
     document.querySelector('.file-name').textContent = fileNameAndSize;
-})
+});
 
 // ----------- Spinner -----------
 function showSpinner() {
     const spinner = document.getElementById('spinner');
-    spinner.classList.replace('spinner-off', 'lds-ring')
+    spinner.classList.replace('spinner-off', 'lds-ring');
     const body = document.querySelector('body');
     body.classList.add('on');
-}
+};
 function hideSpinner() {
     const spinner = document.getElementById('spinner');
     spinner.classList.replace('lds-ring', 'spinner-off');
     const body = document.querySelector('body');
     body.classList.remove('on');
-}
+};
 
 // ----------- DXF -----------
 function uploadDXF() {
     const formData = new FormData();
     const file = document.getElementById("file").files[0];
-    formData.append('dxf', file)
+    formData.append('dxf', file);
     const myInit = {
         method: "POST",
         body: formData,
         headers: {
             "enctype": "multipart/form-data"
         },
-    }
+    };
     fetch("http://localhost:3000/api/mylaser/dxf", myInit)
-    .then((res) => res.json())
-    .then((arrayData) => {
-        hideSpinner();
-        showSVG(arrayData);
-        showOptions();
-    })
-}
+    .then(res => {
+        if(!res.ok) {
+            hideSpinner();
+            document.getElementById('box-error').classList.replace('box-hidden', 'box-visible');
+        } else {
+            res.json()
+            .then((arrayData) => {
+                hideSpinner();
+                showSVG(arrayData);
+                showOptions();
+            });
+        };
+    });
+};
 
 // ----------- Show SVG -----------
 function showSVG(arrayData) {
     // Load SVG in DOM
     let svgContainer = document.getElementById('svgContainer');
     svgContainer.innerHTML = arrayData[0];
-    let svg = document.querySelector('svg')
-    let widthSvg = svg.getAttribute('width')
-    let heightSvg = svg.getAttribute('height')
-    document.getElementById('hauteur').innerHTML = Math.round(heightSvg)
-    document.getElementById('largeur').innerHTML = Math.round(widthSvg)
-    let surfaceUtile = heightSvg*widthSvg
+    let svg = document.querySelector('svg');
+    let widthSvg = svg.getAttribute('width');
+    let heightSvg = svg.getAttribute('height');
+    document.getElementById('hauteur').innerHTML = Math.round(heightSvg);
+    document.getElementById('largeur').innerHTML = Math.round(widthSvg);
+    let surfaceUtile = heightSvg*widthSvg;
     // Paths
     let lengthPath = [];
     let paths = document.querySelectorAll('path');
     paths.forEach(path => {
         let Dpath = path.getAttribute('d');
         lengthPath.push(SVGPathCommander.getTotalLength(Dpath));
-        path.removeAttribute('style')
-        path.setAttribute('style', 'stroke:white;stroke-width:1')
+        path.removeAttribute('style');
+        path.setAttribute('style', 'stroke:white;stroke-width:1');
     });
     // Circles
     let circles = document.querySelectorAll('circle');
@@ -78,9 +85,9 @@ function showSVG(arrayData) {
         let pathCircle = SVGPathCommander.shapeToPath(circle);
         let Cpath = pathCircle.getAttribute('d');
         lengthPath.push(SVGPathCommander.getTotalLength(Cpath));
-        circle.removeAttribute('style')
-        circle.setAttribute('style', 'stroke:white;stroke-width:1')
-    })
+        circle.removeAttribute('style');
+        circle.setAttribute('style', 'stroke:white;stroke-width:1');
+    });
     // Lines
     let lines = document.querySelectorAll('line');
     lines.forEach(line => {
@@ -98,28 +105,28 @@ function showSVG(arrayData) {
         let linePath = SVGPathCommander.shapeToPath(lineAttributes);
         let Lpath = linePath.getAttribute('d');
         lengthPath.push(SVGPathCommander.getTotalLength(Lpath));
-        line.removeAttribute('style')
-        line.setAttribute('style', 'stroke:white;stroke-width:1')
-    })
+        line.removeAttribute('style');
+        line.setAttribute('style', 'stroke:white;stroke-width:1');
+    });
     // Total Paths
     let totalLengthPath = 0;
     for (let i = 0; i < lengthPath.length; i++) {
         totalLengthPath += lengthPath[i];
-    }
+    };
     // Coef Acier
-    let coef =  parseFloat(document.getElementById('width-options').value)
+    let coef =  parseFloat(document.getElementById('width-options').value);
     // Quantity
-    let quantity = parseInt(document.getElementById('quantityList').value)
+    let quantity = parseInt(document.getElementById('quantityList').value);
     // Steel
-    let steel = document.getElementById('steel-options').value
+    let steel = document.getElementById('steel-options').value;
     // Thickness
-    let thickness = document.getElementById('width-options').value
+    let thickness = document.getElementById('width-options').value;
     // Scale SVG for frontend
-    svg.removeAttribute('width')
-    svg.setAttribute('width', '100%')
-    svg.removeAttribute('height')
-    svg.setAttribute('height', '100%')
-    let svgData = JSON.stringify(arrayData[0])
+    svg.removeAttribute('width');
+    svg.setAttribute('width', '100%');
+    svg.removeAttribute('height');
+    svg.setAttribute('height', '100%');
+    let svgData = JSON.stringify(arrayData[0]);
     // Send Quote to DB
     const quoteInfos = {
         length: totalLengthPath,
@@ -132,41 +139,67 @@ function showSVG(arrayData) {
         svg: svgData,
         steel: steel,
         thickness: thickness
-    }
+    };
     const myInit2 = {
         method: "POST",
         body: JSON.stringify(quoteInfos),
         headers: {
             "Content-Type": "application/json; charset=utf-8"
         },
-    }
+    };
     fetch("http://localhost:3000/api/mylaser/dxf/quote", myInit2)
-    .then((res) => res.json())
-    .then((quote) => {
-        localStorage.setItem('currentQuote', JSON.stringify(quote.id))
-        document.getElementById('price').textContent = (quote.price).toFixed(2)
-    })
-}
+    .then(res => {
+        if(!res.ok) {
+            // Error states
+            console.log(arrayData[1])
+            const fileDXF = {
+                filename: arrayData[1]
+            };
+            const myInitFile = {
+                method: "DELETE",
+                body: JSON.stringify(fileDXF),
+                headers: {
+                    "Content-Type": "application/json; charset=utf-8"
+                },
+            };
+            console.log(myInitFile)
+            fetch("http://localhost:3000/api/mylaser/dxf/file", myInitFile)
+            .then(() => {
+                document.getElementById('box-error').classList.replace('box-hidden', 'box-visible');
+                document.getElementById('infos-dxf').classList.replace("visible", "hidden");
+            });
+        } else {
+            res.json()
+            .then((quote) => {
+                localStorage.setItem('currentQuote', JSON.stringify(quote.id))
+                document.getElementById('price').textContent = (quote.price).toFixed(2)
+            })
+            .catch(function (error) {
+                console.log(error)
+            });
+        };
+    });
+};
 
 function showOptions() {
-    document.getElementById('infos-dxf').classList.replace("hidden", "visible")
-}
+    document.getElementById('infos-dxf').classList.replace("hidden", "visible");
+};
 
-widthOptions()
+widthOptions();
 
 function updatePrice() {
-    const currentQuote = localStorage.getItem('currentQuote')
-    const coef = document.getElementById('width-options').value
-    const quantity = document.getElementById('quantityList').value
-    const thickness = document.getElementById('width-options').value
-    const steel = document.getElementById('steel-options').value
+    const currentQuote = localStorage.getItem('currentQuote');
+    const coef = document.getElementById('width-options').value;
+    const quantity = document.getElementById('quantityList').value;
+    const thickness = document.getElementById('width-options').value;
+    const steel = document.getElementById('steel-options').value;
 
     const modifs = {
         coef: coef,
         quantity: quantity,
         steel: steel,
         thickness: thickness
-    }
+    };
 
     const myInit = {
         method: "PUT",
@@ -174,13 +207,13 @@ function updatePrice() {
         headers: {
             "Content-Type": "application/json; charset=utf-8"
         },
-    }
+    };
     fetch(`http://localhost:3000/api/mylaser/dxf/quote/${currentQuote}`, myInit)
     .then((res) => res.json())
     .then((json) => {
-        document.getElementById('price').textContent = (json.price).toFixed(2)
-    })
-}
+        document.getElementById('price').textContent = (json.price).toFixed(2);
+    });
+};
 
 function widthOptions() {
     let steelValue = document.getElementById('steel-options').value;
@@ -213,68 +246,74 @@ function widthOptions() {
         case 'Alu brillant' : 
         widthOption.innerHTML = `<option class="options-width" value="1.3">1.3 MM</option><option class="options-width" value="1.7">1.7 MM</option><option class="options-width" value="1.9">1.9 MM</option>`;
         break;
-    }
-}
+    };
+};
 
 // Update price on change Options
 document.getElementById('steel-options').addEventListener('change', (e) => {
-    widthOptions()
-    updatePrice()
-})
+    widthOptions();
+    updatePrice();
+});
 document.getElementById('width-options').addEventListener('change', (e) => {
-    updatePrice()
-})
+    updatePrice();
+});
 document.getElementById('quantityList').addEventListener('change', (e) => {
-    updatePrice()
-})
+    updatePrice();
+});
 // Add to cart
 document.getElementById('addToCart').addEventListener('click', (e) => {
-    addToCart()
-})
+    addToCart();
+});
 // If change file, delete current quote and reset
 document.getElementById('file').addEventListener('change', (e) => {
+    hideError();
+    changeQuote();
     if(document.getElementById('infos-dxf').className === 'visible') {
         deleteQuote();
-        changeQuote();
         hideOptions();
-    }
-})
+    };
+});
 function deleteQuote() {
     const quote = localStorage.getItem('currentQuote');
-    fetch(`http://localhost:3000/api/mylaser/dxf/quote/${quote}`, {method: "DELETE"})
-    .then(() => localStorage.removeItem('currentQuote'))
-}
+    if(quote) {
+        fetch(`http://localhost:3000/api/mylaser/dxf/quote/${quote}`, {method: "DELETE"})
+        .then(() => localStorage.removeItem('currentQuote'));
+    };
+};
 function changeQuote() {
     let svgContainer = document.getElementById('svgContainer');
     svgContainer.innerHTML = '';
-}
+};
 function hideOptions() {
-    document.getElementById('infos-dxf').classList.replace("visible", "hidden")
-}
+    document.getElementById('infos-dxf').classList.replace("visible", "hidden");
+};
+function hideError() {
+    document.getElementById('box-error').classList.replace('box-visible', 'box-hidden');
+};
 
 function addToCart () {
     const currentQuote = localStorage.getItem('currentQuote');
-    console.log(currentQuote)
+    console.log(currentQuote);
     fetch(`http://localhost:3000/api/mylaser/dxf/quote/${currentQuote}`)
     .then((res) => res.json())
     .then((json) => {
         let currentCart = JSON.parse(localStorage.getItem('currentCart'));
-        console.log(currentCart)
+        console.log(currentCart);
         if(currentCart !== null) {
             if(currentCart[json.id] === undefined) {
-                currentCart.push(json.id)
-            }
-            console.log('existe')
+                currentCart.push(json.id);
+            };
+            console.log('existe');
         } else {
-            console.log('n existe pas')
-            currentCart = [json.id]
-        }
-        localStorage.setItem('currentCart', JSON.stringify(currentCart))
+            console.log('n existe pas');
+            currentCart = [json.id];
+        };
+        localStorage.setItem('currentCart', JSON.stringify(currentCart));
     })
     .then(() => {
         apparitionNext();
-    })
-}
+    });
+};
 
 function apparitionNext () {
     const box = document.createElement('span');
@@ -291,13 +330,13 @@ function apparitionNext () {
     main.classList.add('on2');
 
     document.getElementById('go-to-import').addEventListener('click', () => {
-        window.location.href = '/importdxf.html'
-    })
+        window.location.href = '/importdxf.html';
+    });
 
     document.getElementById('go-to-cart').addEventListener('click', () => {
-        window.location.href = '/cart.html'
-    })
-}
+        window.location.href = '/cart.html';
+    });
+};
